refactor schedulers
