+++ Algorithm-2.0-master/Preferences.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 
 namespace Scheduler {
-    class Preferences {
+    public class Preferences {
         #region NOTES
         //------------------------------------------------------------------------------
         // This Class utilizes a JSON string generated by a SQL query from ParameterSet
