+++ Algorithm-2.0-master/Scheduler.csproj
@@ -298,6 +298,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Algorithms\LongestPathScheduler.cs" />
+    <Compile Include="Algorithms\SchedulerBase.cs" />
     <Compile Include="Contracts\IScheduler.cs" />
     <Compile Include="Contracts\Schedule.cs" />
     <Compile Include="CourseNetwork.cs">
